+++ R/edit_project.R
@@ -67,7 +67,7 @@ edit_project <- function(input = NULL) {
   # Define editable fields per setup function
   config_map <- list(
     "General" = list(setup_fn = setup_project_metadata, prefix = "metadata/", fields = c(
-      "project_name", "project_directory", "dicom_directory", "bids_directory", "fmriprep_directory", "scratch_directory", "templateflow_home",
+      "project_name", "project_directory", "dicom_directory", "bids_directory", "fmriprep_directory", "log_directory", "scratch_directory", "templateflow_home",
       "flywheel_temp_directory", "flywheel_sync_directory", "rois_directory"
     )),
     "Compute Environment" = list(setup_fn = setup_compute_environment, prefix = "compute_environment/", fields = c(